postgresql 8.3 changes
git-svn-id: http://svn.code.sf.net/p/postbooks/code/xtuple/trunk@1931 ea1a15ef-c238-0410-93cb-fa28e9f07733

+++ guiclient/creditcardprocessor.cpp
@@ -816,7 +816,7 @@ int CreditCardProcessor::credit(const int pccardid, const int pcvv, const double
 		"       AS next_seq "
 		"  FROM ccpay "
 		" WHERE (ccpay_order_number=:ccpay_order_number);");
-    ccq.bindValue(":ccpay_order_number", preforder.toInt());
+    ccq.bindValue(":ccpay_order_number", preforder);
     ccq.exec();
     if (ccq.first())
       next_seq = ccq.value("next_seq").toInt();
